+++ lib/angel_static.dart
@@ -1,7 +1,4 @@
 library angel_static;
 
 export 'src/cache.dart';
-export 'src/file_info.dart';
-export 'src/file_transformer.dart';
-export 'src/serve_static.dart';
 export 'src/virtual_directory.dart';