+++ test/cache_sample.dart
@@ -1,24 +1,25 @@
-import 'dart:io';
 import 'package:angel_framework/angel_framework.dart';
 import 'package:angel_static/angel_static.dart';
+import 'package:file/file.dart';
+import 'package:file/local.dart';
 
 main() async {
   Angel app;
-  Directory testDir = new Directory('test');
-  app = new Angel(debug: true);
+  Directory testDir = const LocalFileSystem().directory('test');
+  app = new Angel();
 
-  await app.configure(new CachingVirtualDirectory(
-      source: testDir,
-      maxAge: 350,
-      onlyInProduction: false,
-      // useWeakEtags: false,
-      //publicPath: '/virtual',
-      indexFileNames: ['index.txt']));
+  app.use(
+    new CachingVirtualDirectory(app, const LocalFileSystem(),
+        source: testDir,
+        maxAge: 350,
+        onlyInProduction: false,
+        indexFileNames: ['index.txt']).handleRequest,
+  );
 
   app.get('*', 'Fallback');
 
   app.dumpTree(showMatchers: true);
 
-  await app.startServer(InternetAddress.LOOPBACK_IP_V4, 0);
-  print('Open at http://${app.httpServer.address.host}:${app.httpServer.port}');
+  var server = await app.startServer();
+  print('Open at http://${server.address.host}:${server.port}');
 }