+++ test/cache_test.dart
@@ -1,30 +1,33 @@
-import 'dart:io';
 import 'package:angel_framework/angel_framework.dart';
 import 'package:angel_static/angel_static.dart';
+import 'package:file/file.dart';
+import 'package:file/local.dart';
 import 'package:http/http.dart' show Client;
 import 'package:matcher/matcher.dart';
 import 'package:test/test.dart';
 
 main() {
   Angel app;
-  Directory testDir = new Directory('test');
+  Directory testDir = const LocalFileSystem().directory('test');
   String url;
   Client client = new Client();
 
   setUp(() async {
-    app = new Angel(debug: true);
+    app = new Angel();
 
-    await app.configure(new CachingVirtualDirectory(
-        source: testDir, maxAge: 350, onlyInProduction: false,
-        //publicPath: '/virtual',
-        indexFileNames: ['index.txt']));
+    app.use(
+      new CachingVirtualDirectory(app, const LocalFileSystem(),
+          source: testDir, maxAge: 350, onlyInProduction: false,
+          //publicPath: '/virtual',
+          indexFileNames: ['index.txt']).handleRequest,
+    );
 
     app.get('*', 'Fallback');
 
     app.dumpTree(showMatchers: true);
 
-    await app.startServer(InternetAddress.LOOPBACK_IP_V4, 0);
-    url = "http://${app.httpServer.address.host}:${app.httpServer.port}";
+    var server = await app.startServer();
+    url = "http://${server.address.host}:${server.port}";
   });
 
   tearDown(() async {
@@ -40,19 +43,14 @@ main() {
 
     expect(response.statusCode, equals(200));
     expect(
-        [
-          HttpHeaders.ETAG,
-          HttpHeaders.CACHE_CONTROL,
-          HttpHeaders.EXPIRES,
-          HttpHeaders.LAST_MODIFIED
-        ],
+        ['ETag', 'cache-control', 'expires', 'last-modified'],
         everyElement(predicate(
             response.headers.containsKey, 'contained in response headers')));
   });
 
   test('if-modified-since', () async {
     var response = await client.get("$url", headers: {
-      HttpHeaders.IF_MODIFIED_SINCE:
+      'if-modified-since':
           formatDateForHttp(new DateTime.now().add(new Duration(days: 365)))
     });
 
@@ -60,11 +58,7 @@ main() {
 
     expect(response.statusCode, equals(304));
     expect(
-        [
-          HttpHeaders.CACHE_CONTROL,
-          HttpHeaders.EXPIRES,
-          HttpHeaders.LAST_MODIFIED
-        ],
+        ['cache-control', 'expires', 'last-modified'],
         everyElement(predicate(
             response.headers.containsKey, 'contained in response headers')));
   });